requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pwdlib>=0.2.0
pyjwt>=2.8.0